BFS pathfinding algo based off senior assumption
According to seniors, the system uses a grid, with obstacles taking up a 1 by 1 square in the grid and the car taking up 2 by 2. The camera is positioned at the side and takes up a 1 by 1 square of the 2 by 2 car. The task is to take an image of the obstacles at one side. If the car can only make left turns, right turns, move forward, and move backward, there are 3 moves that it can make to face the desired side (turning, or moving straight). This means that for each desired side there are 3 goal squares, and it has to reach one of the three to take a picture.

In sequence, the car will first navigate to the goal square (square 1 move away to the desired obstacle face).
It will then turn to face the desired side and take a photo.
It will then reverse that turn back to its original position.
It will continue on to the next goal square of the next obstacle.

The BFS algo works off these assumptions, is yet to be integrated and is currently hardcoded for now. It allows you to set obstacles and goals, and finds the shortest path to all the goals in the shortest sequence without backtracking, or hitting any defined obstacle. It spits out the correct sequence of goals, and the commands (turn left, turn right, go straight etc) required to achieve that in reverse order. Further improvements include coming up with the instructions in proper order, and including turn and reverse turn the car will have to make upon reaching a goal square.